pyproject.toml
[project]
name = "md2img"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "markdown[extra]>=3.10.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
